webapps/requirements.txt
dj-database-url==0.4.1
Django==1.10.1
gunicorn==19.6.0
Pillow==3.2.0
psycopg2==2.6.2
whitenoise==3.2.2
pytz==2016.4
django-storages==1.5.1
boto==2.43.0
channels==0.17.3
django-vote==2.1.2
asgi-redis==1.0.0
asgiref==1.0.0
daphne==0.15.0
redis==2.10.5
